Move build scrip to a gulp task
Remove `build-types` and `build-js` package scripts and instead export only a single gulp task called "build-js" and bind that to a package script called "build". Moving this work into gulp tasks will be required when we add any watchers in the future needed for a live pattern lib or express server.

Copy run.js from the client and move it into a new exported scripts/ folder that can be used for common scripts between frontend packages.