Add HTMLAttributes to button props
The button should be able to set the `type` prop . Otherwise it defaults to "submit" and will submit any form that embeds it. `type` is a native prop on button and so it makes sense to just inherit the actual types from button from Preact. In doing so, we can remove some redundant props and replace them with HTMLInputElement props.

`size` is the one prop that conflicts because of the way Preact constructs HTMLInputElement--which is derived from HTMLElement and inherits the `size` prop even though it's not relevant for button. The simple fix is to omit `size` when pulling in HTMLInputElement.