Make "Example" layout columnar for narrower screens
The newer `Example` layout for pattern-library examples is more
conducive to responsive-ness than the older `PatternExamples` and
`PatternExample` styling. Make it columnar on narrower screens.

`Example` is still a prototype layout, but may supersede/replace the
table-based `PatternExample(s)` styling in the future.

The `Example` layout is used on the `Dialogs` component pattern library
page, and it needs to be responsive so one can view the responsive
modal patterns in a responsive setting.