Add icons to package
Make shared icons available in the package entrypoint.

This will allow applications that use this package to merge in the
shared icons exported here with their own local icons for ultimate
registration.

This will prevent applications from breaking if the
contents (e.g. components) in this package are updated to use new icons
that are not registered in the applications that use this package.

Applications will need to be updated to merge in this icon set to their
own icons for registration.