Implement SVG-styled Checkbox
Add an SVG image directly after checkbox input elements that can be
used to style the appearance of the checkbox itself. The checkbox
input element is still what the user interacts with, composited
visually with an SVG. This gives us the behavior of the underlying
<input> element but allows us to visually style the checkbox as we
desire.

See https://www.sarasoueidan.com/blog/inclusively-hiding-and-styling-checkboxes-and-radio-buttons/