Add CSS to horizontally-center responsive `IconButton`s
There is a media-query that will enforce a minimum width and height on
`IconButton`s to ensure a minimum tap-target size. Before these changes,
icons within `IconButton`s on narrow screens could appear off-center
horizontally.

The icon image is already vertically centered.
Add a horizontal-alignment rule here to ensure that the icon image
always appears in the center of the entire button.