Revise icon registration API to prevent conflicts
Revise the API used to register icons with `SvgIcon` to prevent
conflicts between different packages using the component.

Previously the keys of registered icons were developer-chosen strings.
This could lead to conflicts if different packages tried to use the
same string for different icons. The new approach in this commit is for
the developer to call a `registerIcon(name, svgMarkup)` function which
registers an icon and returns a key for use with `<SvgIcon name={key}
..>`. `registerIcon(...)` creates a new symbol and uses it as the key
for the map entry, then returns the key. It could just as well generate
some other type of unique key, but a symbol seems the natural choice for
this purpose.

Previous API:

```js
registerIcons({ 'cancel': require('path/to/icon.svg') })

<SvgIcon name="cancel" .../>
```

New API:

```
const cancelIcon = registerIcon('cancel', require('path/to/icon.svg'))

<SvgIcon name={cancelIcon} .../>
```

To enable gradual migration, the `registerIcons` functions and support
for string values for the `name` prop is maintained, but deprecated.