Dual-build package as ES modules and CommonJS
This will smooth the migration of downstream applications to a new
bundler that speaks ES modules natively [1]. The existing `lib/` dir and
`.babelrc` module now relate to the ES module build, and the `lib-cjs/`
dir and `.babelrc-cjs` module are for the CommonJS build. Once all
downstream dependencies have been switched over to CommonJS we can
remove these.

In the process I changed the file extension of the Babel config files from .cjs
to .js which results in them being formatted by Prettier like other .js
files.

The browser targets of the `@babel/preset-env` package have also been
updated to reflect the current reality [2] of what we support.

[1] In particular it is important that Preact is only included once in
    the generated app bundle. For that to happen both the app and
    @hypothesis/frontend-shared need to reference the same build of Preact
    (ES modules or CommonJS).
[2] https://github.com/hypothesis/client/issues/3744