Remove the CommonJS build
All of the downstream projects have converted to use Rollup, so we no
longer need to publish a CommonJS build of this package.

One way to do this is to mark the entire package as being an ESM
package by setting `"type": "module"` in package.json, indicating that
the "main" field refers to an ES module.

As a result, all .js files in the package are interpreted as ESM, unless
they are renamed to have a .cjs extension.

 - Remove the CommonJS build steps and files from the package

 - Set `"type": "module"` in package.json and make the `main` field
   point to the ESM files

 - Convert scripts/serve-pattern-library.js and .babelrc.js to be ES modules

 - Rename src/karma.config.js to have a .cjs extension, because Karma
   still expects to load it as a module

 - Change .mjs files to just have a .js extension again, since the whole
   package is now ESM.