Fix layout of demo source in pattern-library examples
The recent removal of all reset from pattern-library styles caused a
visual regression in the styling of the "Source" tab in pattern-library
demos.

Fix this, and add a couple of (perhaps-temporary, we'll see) reset mixins
based on some existing things in the `client` app. These mixins are
only used locally currently.