Add custom tailwind variant for applying clean-theme styling
Add a custom `theme-clean:` tailwind modifier (variant) to the preset
such that it is possible to apply styles for the clean theme to any
element.

E.g. `theme-clean:bg-white` would apply a white background color to an
element ONLY if it was in a document with a parent element that had the
`.theme-clean` class applied.

https://tailwindcss.com/docs/plugins#adding-variants