Intercept in-app anchor links and trigger client-side navigation
Previously clicking on a link from within the content of a page to another page
would trigger a full page reload. Only the links in the navigation bar used
client-side navigation. Update the router to intercept all clicks on "in-app"
URLs and handle them client-side.

In the process add missing types in `router.js`, and remove the custom `onClick`
handlers for links in the navbar, which are no longer required.

We discussed that we might want to replace the custom router with a 3rd-party
solution at some point, but we will need to evaluate options.