Fix conflict between Overlay's `open` prop and JSX "open" attribute
Both Overlay and the HTML details element accept an `open` prop. This used to be
`boolean|undefined` in both cases, so there was no conflict. In the latest
Preact update however all HTML attributes have gained the ability to accept
signals as well as raw values. This means the types of Overlay's `open` and JSX
`open` no longer align.