Update release process and docs to work with GH rebase-and-merge method
If we want to use `rebase-and-merge` as our GitHub merge method, we
need to adjust our release process to ensure that the version tag points
at the intended commit. Rebase-and-merge rewrites commit hashes, so tags
should not be created until after the target commit is present on the
`main` branch.