Clarify types for pattern-library entry point and app component
Consolidate type definitions and comments pertaining to the pattern
library entry point and its outermost app component for clarity and to
avoid redundancy.

Typically we define types for a component's props in the component
module, but in this case, props for the pattern library app component
are typed in the app's entry point for easier reference.