Fix default `baseURL` in pattern library
Previously, both the entry point and `PlaygroundApp` set defaults for
the same prop, `baseURL`, and they were setting a different default.
Now only `PlaygroundApp` sets this default, and it should be `''`, not
`/ui-playground/`.

This fixes routing such that reloading works and the router-generated
URL is correct (in this project's pattern library).