Adjust common tests to allow distinction of composite wrapper selector
The common tests already take an `elementSelector` option to allow for
components whose primary element is not the outermost element rendered
by the component.

This change adds a `wrapperSelector` option for composite element tests
to allow for cases when the wrapping element (i.e. the element to
which the `data-composite-component` attribute is applied) differs from
the primary element (i.e. to which HTML attributes are merged).