Add `focusElement` callback to `useArrowKeyNavigation` hook
This allows users of the hook to customize what happens when the hook tries to
focus an element, and also get the context of the `KeyboardEvent` which
triggered the focus, if any.

This will be needed in `DataTable` where we want to know which keyboard
modifiers (shift, ctrl etc.) were pressed.