Add `elementRef` prop to `Popover` component
This is a ref for the popover's outermost element, like `elementRef` for other
components. This is useful for downstream code which needs information about the
geometry of the rendered popover.  Downstream code could use a ref on the
popover's content instead, but that doesn't include any padding which the
popover has.