Add `lazy` utility to create lazy-loading UI components
Add a utility for creating components which are loaded only when rendered. When
first rendered, a lazy component displays a fallback while the real
implementation is fetched. It then renders the real component after the fetch
completes. If the component fails to load an an error fallback is rendered in
its place.