Prevent checkbox / radio button icon from being clipped when label is long
When a checkbox or radio button had a label that filled the entire available
width, the icon would begin to shrink and would be clipped. The icon had a
preferred size set, but not a minimum size.

Set the minimum size to match the SVG, which means 16px.

Part of a fix for https://github.com/hypothesis/h/issues/9635.