Replace focus-ring-visible usage with Tailwind's built-in equivalent
Replace uses of our custom `focus-ring-visible` utility with Tailwind's built-in
equivalent styles (`focus-ring:visible focus-ring:outline-none`). The original
reason for our custom utility was to support a polyfill for `:focus-visible`,
but we no longer need/use that polyfill.

This looks the same as before but will be easier to migrate to Tailwind v4. In
Tailwind v3 `outline-none` doesn't actually generate `outline-style: none`, but
the result is equivalent for our purposes.  See
https://tailwindcss.com/docs/upgrade-guide#renamed-outline-utility.

The focus-ring-visible plugin is still included in the package because
downstream apps are still using it. We can remove it after the same migration is
applied there.